refactor process mgmt and split Pass into subclasses
