+++ qtpass.pro
@@ -34,7 +34,10 @@ SOURCES   += main.cpp\
              qprogressindicator.cpp \
              qpushbuttonwithclipboard.cpp \
              qtpasssettings.cpp \
-             settingsconstants.cpp
+             settingsconstants.cpp \
+             pass.cpp \
+			 realpass.cpp \
+			 imitatepass.cpp
 
 HEADERS   += mainwindow.h \
              configdialog.h \
@@ -50,9 +53,9 @@ HEADERS   += mainwindow.h \
              qtpasssettings.h \
              enums.h \
              settingsconstants.h \
-             pass.h \
-             realpass.h \
-             imitatepass.h
+			 pass.h \
+			 realpass.h \
+			 imitatepass.h
 
 FORMS     += mainwindow.ui \
              configdialog.ui \
